Sending data between pico and ev3
We can now communicate between the ev3 and the pi pico. The ev3 sends the character 'd' and the pico responds with teh curreent angle. We also did an fft plot to see noise in the gyro data

+++ pico_gyro/mpu6050.h
@@ -13,8 +13,8 @@
 #define MPU_ADDR 0x68
 
 typedef struct {
-  int16_t accel_offset[3];
-  int16_t gyro_offset[3];
+  float accel_offset[3];
+  float gyro_offset[3];
 } mpu6050_calibration_t;
 
 void mpu6050_init();
